new: added Makefile generation
When creating a project, generate a Makefile with targets for the
- linter/formatter
- type checker
- test framework
so that the user can use the same commands uniformly across frameworks.

For example, a user wouldn't need to know if a krait-generated
project used `unittest`, `pytest`, `nose`, etc. -- the user need only
type `make test` at the command prompt and the generated target will
perform the appropriate action.